Reduce prettier overrides (#30)
## Ticket

#24 

## Changes
This PR removed some non-default settings for prettier and then also
made the formatting changes associated with that setting change. The
setting change and the format change are in separate commits, in case
that makes it easier to review.

## Context for reviewers
All changes should be non-functional, so just visual confirmation that
this looks right is probably all that's needed. Rationale for the change
is provided in the issue (#24).

## Testing
Non-functional changes, so little testing was needed, but I still ran
storybook and the next.js app to make sure those still work, and they
did.