Readme updates to sync with `package.json` (#50)
## Ticket

resolves #23

## Changes
This changeset is only updating the documentation of the `package.json`
file in the Readme.

## Context for reviewers
I added documentation for all of the dependencies in scripts that are in
`package.json`. I'm not super experienced with frontend repos, so some
of my terminology and descriptions may be imprecise, I'm very open to
suggestions and feedback on that.

## Testing
I previewed the Markdown and it looks as expected.