Refactor README's (#67)
## Changes

- Rewrite the root `README` and `app/README` so that:
- Root `README` introduces viewers to the _template_ and how they might
use the template for their own project. The file can be removed after
the project repo is created.
- `app/README` focuses on development instructions. It's written from
the perspective of a running a project, so that projects don't have to
update this to make sense for their team (i.e removing template
directions or explainations).
- Restructured `app/README` around topic (Development, Storybook, Tests,
Type checking) and tried to make it more succinct/scannable (kinda
subjective tho!).
- Fixed issue where the example API route wasn't working because it
needed nested in the `pages/` directory
- Removed the `@pages` alias route since it was unused (and I think
would have required additional config for support in tests and
Storybook).