Upgrade to Next.js 13 (#69)
## Ticket

Closes #65

## Context for reviewers

Next.js 13 introduced [pretty significant new
functionality](https://nextjs.org/blog/next-13), but it's backwards
compatible with the existing approach implemented in this repo, so
should be safe to upgrade. Projects can adopt its new functionality as
they want. We have a separate ticket for adopting the new functionality
in this repo: #66