Redesign the default landing page (#145)
## Changes

This redesigns the landing page, with the goal of helping teams more
quickly create a more polished looking site, and to demonstrate common
needs like responsive utility classes and translated strings with HTML
formatting:

- Adds the USWDS header and footer components from the `react-uswds`
library. Included is "sticky footer" behavior, so that the footer is
always at the bottom of the viewport, even when the page body doesn't
extend to the bottom.
- Adds body content to demonstrate usage of the `Trans` component
- Removed the `react-uswds-hello.tsx` page, since there are now other
files demonstrating usage of those React USWDS components

## Context for reviewers

After observing a few more projects using this template, and using it a
few more times myself, I've observed:

1. The USWDS header doesn't seem to support a logo image out of the box,
so there's some custom code needed for nice visual alignment and mobile
responsiveness. This can take a few hours and is confusing.
2. Rendering translated strings with HTML formatting trips up folks,
particularly because there are several ways of doing it in `i18next`.
3. Sticking the footer to the bottom of the page doesn't get identified
as a thing that needs to happen, or doesn't get prioritized right away.

## Testing

<img width="959" alt="CleanShot 2023-05-08 at 11 38 58@2x"
src="https://user-images.githubusercontent.com/371943/236904761-75a76b6f-3ca6-4228-891a-a5f2415d9041.png">