Use next-intl for internationalization, replacing i18next (#260)
## Ticket

Part of #66 

## Changes

- Replace all I18next dependencies with
[`next-intl`](https://next-intl-docs.vercel.app/)
- Add new `tests/test-utils` file that would be used for rendering and
querying a React tree in tests, rather than directly importing from
`@testing-library/react`. This is so that the i18n content is provided
to the component being tested.

## Context for reviewers

This is one of the main pieces of the larger migration effort from the
Pages router to the App Router (#66). To reduce the size of that PR,
I've broken out the i18n migration work into this PR.

Next.js App Router doesn't support internationalized routing or locale
detection, like Pages router, and `next-intl` makes it easy to restore
that functionality via middleware.