Dynamically load locale messages (#261)
## Ticket

#66 

## Changes

- Use a [dynamic
import](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/import)
to load all strings for the selected locale. This simplifies the process
of adding a new language — now someone just needs to add the locale code
(e.g. `es-US`) to the `locales` array, whereas before they'd also have
to add an import statement.
- Renamed `getLocaleMessages` to `getMessagesWithFallbacks` and moved it
to its own file. A few reasons for this:
(1) Project engineers shouldn't ever need to interact with the
`getMessagesWithFallbacks` logic — it should Just Work. (2) We can now
rename `i18n/index.ts` to `i18n/config.ts` to make it clearer what the
purpose of that file is.

## Context for reviewers

I'm pulling this out from the larger App Router migration branch.
There's an additional piece that will be in that branch, resulting in a
final directory structure looking something like this:

```
├── i18n
│   ├── config.ts     # Supported locales and formatting options
│   ├── getMessagesWithFallbacks.ts 
│   └── server.ts     # 🆕 next-intl setup logic that can't be imported into client components
```

We could expand this also to export the safelisted subset of
hooks and components

```
├── i18n
│   ├── config.ts     # Supported locales and formatting options
│   ├── getMessagesWithFallbacks.ts 
│   ├── index.ts      # 🆕 Hooks and components for rendering content
│   └── server.ts     # next-intl setup logic that can't be imported into client components
```