Don't set default port in Dockerfile
Doesn't _really_ matter, but it's not necessary and could be confusing
when the actual configured port locally or hosted could be wildly
different.

And in general the production build shouldn't have settings (even just
defaults) burnt into the image. Runtime configuration should be
configured in the runtime, or fallback to defaults in the code itself
which is more obvious.